+++ flutter_libs/windows/include/objectbox_flutter_libs/objectbox_flutter_libs_plugin_c_api.h
@@ -1,5 +1,5 @@
-#ifndef FLUTTER_PLUGIN_OBJECTBOX_FLUTTER_LIBS_PLUGIN_H_
-#define FLUTTER_PLUGIN_OBJECTBOX_FLUTTER_LIBS_PLUGIN_H_
+#ifndef FLUTTER_PLUGIN_OBJECTBOX_FLUTTER_LIBS_PLUGIN_C_API_H_
+#define FLUTTER_PLUGIN_OBJECTBOX_FLUTTER_LIBS_PLUGIN_C_API_H_
 
 #include <flutter_plugin_registrar.h>
 
@@ -13,11 +13,11 @@
 extern "C" {
 #endif
 
-FLUTTER_PLUGIN_EXPORT void ObjectboxFlutterLibsPluginRegisterWithRegistrar(
+FLUTTER_PLUGIN_EXPORT void ObjectboxFlutterLibsPluginCApiRegisterWithRegistrar(
     FlutterDesktopPluginRegistrarRef registrar);
 
 #if defined(__cplusplus)
 }  // extern "C"
 #endif
 
-#endif  // FLUTTER_PLUGIN_OBJECTBOX_FLUTTER_LIBS_PLUGIN_H_
+#endif  // FLUTTER_PLUGIN_OBJECTBOX_FLUTTER_LIBS_PLUGIN_C_API_H_