+++ objectbox/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## latest
 
+* Requires at least Dart SDK 3.12 or Flutter SDK 3.44.
+  * Android apps: min SDK increased to 24 (Android 7.0).
+* Flutter plugins: support Swift Package Manager [#707](https://github.com/objectbox/objectbox-dart/issues/707) and built-in Kotlin [#812](https://github.com/objectbox/objectbox-dart/issues/812).
+
 ## 5.3.2 (2026-05-20)
 
 * Update ObjectBox database for Flutter Linux/Windows, Dart Native apps to [5.3.2-2026-05-05](https://github.com/objectbox/objectbox-c/releases/tag/v5.3.2)